move netdever funcs to stacks/netdev.go
